Added WHOI Deep Submergence environment models and examples
NOTE: Requires the Woods Hole ds_sim and ds_msgs packages to be
downloaded from the Woods Hole Bitbucket repository and compiled
(ds_sim package must be on the feature/melodic branch to compile
successfully for ROS Melodic).  In order to function, the WHOI
ROS plugins require custom Gazebo sensors that are included in
the ds_sim package.  These sensors are made available to Gazebo
with the -s libdsros_sensors.so option (see launch file examples).