First cut at a (very) simple visual occlusion example
The simple_occlusion_demo.launch will launch a world very similar to the
grabbable bar demo.  The only differences are that there is about 6"
of "silt" on the bottom that will partially occlude the grabbable bar,
and there is a large "mossy cinderblock" added to the scene.

Bottom occlusion due to silt is implemented by shifting the visual model
away from the collision model slightly.  The occlusion by "moss" on the
cinderblock is accomplished by increasing the scale of the cinderblock's
visual mesh.

Sensors that rely on the graphics pipeline (e.g., GPURay, camera, etc.)
will interact with the visual model.  Those that rely on the physics
engine (Ray, manipulators, etc.) will interact with the collision model.