Eliminated "deprecated" warnings with oberon7 and predator xacro
Added the 'hardware_interface/' preface to the EffortJointInterface
contents of the hardwareInterface elements of the transmission macro.

NOTE: This was already in both .xacro files but commented out, so this
change just comments out the previous macro and uncomments this one.
If the original commenting was there for a reason, this change can
be undone (previous commenting for Gazebo 8?)