Changes ISO DVL current profiling implementation (first cut)
Corresponding DVL-specific changes are currently in the nps_dvl_dev
branch of the NPS copy of the WHOI ds_sim repository.

Ocean current world plugin updated to implement stratified current as
Gauss-Markov processes at each layer (as the global current is done).
A Gazebo StratifiedCurrent message is used to publish the varying
stratified current.

The Ocean current ROS plugin is updated to publish both the fixed
stratified current database values (including tidal values) to a
stratified_ocean_current_database topic and the time-varying stratified
current to a stratified_current_velocity topic.

Models, URDF, launch files, and CMake files were all updated accordingly
so that everything still works (as far as I can tell).

A scope error in the GaussMarkovProcess class was also fixed.  The
seed variable to the thread-safe random number generator was declared
locally, so state was not maintained (same "random" number was generated
every time).  It was moved to a private class variable instead.