Generated Xacro versions of the socket and created a demo
Plug and socket plugin updates to clean up the code and to base
the alignment on the world pose of the plug and socket links.  This
should make it easier to add plug/socket pairs to arbitrary models.

The platform model is just a box for now, but a better model will
be incorporated once a mesh is available.