Fix for bug IDed in pre-merge checks
The error was because the joint was being deleted but not
"removed" from the model when the plug was detached from the
socket.  This caused an error on subsequent attempts to
plug into the socket.

Also did some code cleanup and updated the models so that they
are not so bright on most hosts.