Added ROS publication of applied force to flying lead plugin
Force is calculated as the force applied to the plug link by
either the sensor plate link (during plug insertion) or the
gripper links (during plug extraction). Forces are published
as a 3D vector in the plug link frame (since calculated force
is from different links during insertion and extraction, it will
always be positive along the X axis).  Only force in the X
direction is used to determine whether or not the lock or unlock
threshold has been met.